# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "API for an employee management system to determine employees that\
    \ are no longer working on active projects and to understand their qualifications\
    \ and expertise for suitability in other projects"
  version: "1.0.0"
  title: "UBahn API"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /users/{userId}/externalProfiles:
    get:
      tags:
      - "External Profiles"
      description: "Get External Profiles with given user id.\n\n**Security** - Note\
        \ that for non-admin users, this endpoint will only return entities that\n\
        the user has created.        \n"
      operationId: "usersUserIdExternalProfilesGET"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "organizationName"
        in: "query"
        description: "The organization name"
        required: false
        type: "string"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ExternalProfile"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "ExternalProfiles"
    head:
      tags:
      - "External Profiles"
      description: "Get External Profiles with given user id, but only header information\
        \ is returned.\n\n**Security** - Note that for non-admin users, this endpoint\
        \ will only return entities that\nthe user has created.        \n"
      operationId: "usersUserIdExternalProfilesHEAD"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "ExternalProfiles"
    post:
      tags:
      - "External Profiles"
      description: "Create a new External Profile for given user id.\n\n**Security**\
        \ - This endpoint is accessible by all authenticated users.        \n"
      operationId: "usersUserIdExternalProfilesPOST"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/ExternalProfileRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/ExternalProfile"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "ExternalProfiles"
  /users/{userId}/externalProfiles/{organizationId}:
    get:
      tags:
      - "External Profiles"
      description: "Get external profile with given user id and organization id.\n\
        \n**Security** - Note that for non-admin users, this endpoint will only return\
        \ entities that\nthe user has created.        \n"
      operationId: "usersUserIdExternalProfilesOrganizationIdGET"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "organizationId"
        in: "path"
        description: "The organization id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/ExternalProfile"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "ExternalProfiles"
    head:
      tags:
      - "External Profiles"
      description: "Get external profile with given user id and organization id, but\
        \ only header information is returned.\n\n**Security** - Note that for non-admin\
        \ users, this endpoint will only return entities that\nthe user has created.\
        \        \n"
      operationId: "usersUserIdExternalProfilesOrganizationIdHEAD"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "organizationId"
        in: "path"
        description: "The organization id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "ExternalProfiles"
    delete:
      tags:
      - "External Profiles"
      description: "Remove an existing external profile with given user id and organization\
        \ id.\n\n**Security** - Note that this endpoint is only available for admin\
        \ users.  \n"
      operationId: "usersUserIdExternalProfilesOrganizationIdDELETE"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "organizationId"
        in: "path"
        description: "The organization id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "204":
          description: "OK - the request was successful"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "ExternalProfiles"
    patch:
      tags:
      - "External Profiles"
      description: "Update an existing external profile with given user id and organization\
        \ id.\n\n**Security** - Note that for non-admin users, this endpoint will\
        \ only allow updates on entities that the\ncalling user has created.\n"
      operationId: "usersUserIdExternalProfilesOrganizationIdPATCH"
      parameters:
      - name: "userId"
        in: "path"
        description: "The user id"
        required: true
        type: "string"
        format: "UUID"
      - name: "organizationId"
        in: "path"
        description: "The organization id"
        required: true
        type: "string"
        format: "UUID"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/ExternalProfileUpdateRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/ExternalProfile"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "ExternalProfiles"
...
  /organizations:
    get:
      tags:
      - "Organizations"
      description: "Search organizations in the application.  If no results, then\
        \ empty array is returned.\n\n**Security** - Note that for non-admin users,\
        \ this endpoint will only return entities that\nthe user has created.\n"
      operationId: "organizationsGET"
      parameters:
      - name: "name"
        in: "query"
        description: "Filter by organization name"
        required: false
        type: "string"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Organization"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Organizations"
    head:
      tags:
      - "Organizations"
      description: "Retrieve header information for a search operation on organizations\
        \ in the application.\n\n**Security** - Note that for non-admin users, this\
        \ endpoint will only return entities that\nthe user has created.\n"
      operationId: "organizationsHEAD"
      parameters:
      - name: "name"
        in: "query"
        description: "Filter by organization name"
        required: false
        type: "string"
      responses:
        "200":
          description: "success response"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Organizations"
    post:
      tags:
      - "Organizations"
      description: "Create a new Organization.\n\n**Security** - This endpoint is\
        \ accessible by all authenticated users.        \n"
      operationId: "organizationsPOST"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/NameRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/Organization"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Organizations"
  /organizations/{organizationId}:
    get:
      tags:
      - "Organizations"
      description: "Get organization with given id.\n\n**Security** - Note that for\
        \ non-admin users, this endpoint will only return entities that\nthe user\
        \ has created.        \n"
      operationId: "organizationsOrganizationIdGET"
      parameters:
      - name: "organizationId"
        in: "path"
        description: "The organization id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/Organization"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Organizations"
    head:
      tags:
      - "Organizations"
      description: "Get organization with given id, but only header information is\
        \ returned.\n\n**Security** - Note that for non-admin users, this endpoint\
        \ will only return entities that\nthe user has created.        \n"
      operationId: "organizationsOrganizationIdHEAD"
      parameters:
      - name: "organizationId"
        in: "path"
        description: "The organization id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "200":
          description: "OK - the request was successful"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Organizations"
    delete:
      tags:
      - "Organizations"
      description: "Remove an existing organization with given id.\n\n**Security**\
        \ - Note that this endpoint is only available for admin users.  \n"
      operationId: "organizationsOrganizationIdDELETE"
      parameters:
      - name: "organizationId"
        in: "path"
        description: "The organization id"
        required: true
        type: "string"
        format: "UUID"
      responses:
        "204":
          description: "OK - the request was successful"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Organizations"
    patch:
      tags:
      - "Organizations"
      description: "Update an existing organization with given id.\n\n**Security**\
        \ - Note that for non-admin users, this endpoint will only allow updates on\
        \ entities that the\ncalling user has created.\n"
      operationId: "organizationsOrganizationIdPATCH"
      parameters:
      - name: "organizationId"
        in: "path"
        description: "The organization id"
        required: true
        type: "string"
        format: "UUID"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/NameRequestBody"
      responses:
        "200":
          description: "OK - the request was successful"
          schema:
            $ref: "#/definitions/Organization"
        "400":
          $ref: "#/definitions/BadRequest"
        "401":
          $ref: "#/definitions/Unauthorized"
        "403":
          $ref: "#/definitions/Forbidden"
        "404":
          $ref: "#/definitions/NotFound"
        "409":
          $ref: "#/definitions/Conflict"
        "500":
          $ref: "#/definitions/ServerError"
      security:
      - Bearer: []
      x-swagger-router-controller: "Organizations"
...
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  AuditFields:
    type: "object"
    required:
    - "created"
    - "createdBy"
    - "updated"
    - "updatedBy"
    properties:
      created:
        type: "string"
        format: "date-time"
        description: "When the entity was created."
      updated:
        type: "string"
        format: "date-time"
        description: "When the entity was updated."
      createdBy:
        type: "string"
        format: "UUID"
        description: "Creator of the entity."
      updatedBy:
        type: "string"
        format: "UUID"
        description: "User that last updated the entity."
    description: "Describes the audit fields that are present in all the models in\
      \ this API."
...
  ExternalProfile:
    allOf:
    - type: "object"
      required:
      - "organizationId"
      - "uri"
      - "userId"
      properties:
        userId:
          type: "string"
          format: "UUID"
          description: "The id of the user this profile belongs to."
        organizationId:
          type: "string"
          format: "UUID"
          description: "The id of the organization this profile belongs to."
        uri:
          type: "string"
          description: "The uri of the external profile."
    - $ref: "#/definitions/AuditFields"
  ExternalProfileRequestBody:
    type: "object"
    required:
    - "organizationId"
    - "uri"
    properties:
      organizationId:
        type: "string"
        format: "UUID"
        description: "The id of the organization this profile belongs to."
      uri:
        type: "string"
        description: "The uri of the external profile."
    example:
      organizationId: "organizationId"
      uri: "uri"
  ExternalProfileUpdateRequestBody:
    type: "object"
    properties:
      uri:
        type: "string"
        description: "The uri of the external profile."
    example:
      uri: "uri"
...
  Organization:
    allOf:
    - type: "object"
      required:
      - "id"
      - "name"
      properties:
        id:
          type: "string"
          format: "UUID"
          description: "Id of the organization"
        name:
          type: "string"
          description: "Name of the organization"
    - $ref: "#/definitions/AuditFields"
  NameRequestBody:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
        description: "Name of the entity"
    description: "Simple request body containing the name of the entity."
    example:
      name: "name"
...
  Unauthorized:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Unable to authenticate the user."
        description: "The unauthorized error message."
    description: "The unauthorized error entity."
  NotFound:
    type: "object"
    properties:
      message:
        type: "string"
        example: "A resource with the name could not be found."
        description: "The not found error message."
    description: "The not found error entity."
  ServerError:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Something went wrong while processing your request. We�re sorry\
          \ for the trouble. We�ve been notified of the error and will correct it\
          \ as soon as possible. Please try your request again in a moment."
        description: "The server error message."
    description: "The server error entity."
  BadRequest:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Invalid input."
        description: "The bad request error message."
    description: "The bad request error entity."
  Forbidden:
    type: "object"
    properties:
      message:
        type: "string"
        example: "You are not allowed to access the request."
        description: "The forbidden error message."
    description: "The permission error entity."
  Conflict:
    type: "object"
    required:
    - "message"
    properties:
      message:
        type: "string"
        example: "Creating a resource with a name already exists."
        description: "The conflict error message."
    description: "The conflict error entity."